# backend/postman/openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI Documentation","description":"\nAPI endpoints for backend\n---\n\n---\n**Contact**\n\nAuthor: Matías José Varrone\n\nEmail: mativarrone2@gmail.com\n\n---\nMarch, 2022\n","version":"0.0.1"},"paths":{"/create-tax/":{"post":{"tags":["Payable"],"summary":"Create Payable","description":"Create payable\n\n*Parameters in body*\n\n    service_type: (e.g. luz, agua, gas, etc)\n\n    description: company name (e.g. Edenor S.A., Aguas Cordobesas S.A., MetroGAS S.A.)\n\n    due_date: Expiration date (e.g. 2021-01-15). Format: YYYY-MM-DD\n\n    amount: float\n\n    payment_status: \"pending\"\n\n    barcode: unique, 12 digits-length","operationId":"create_payable_create_tax__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Payable"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/unpaid-list/":{"get":{"tags":["Payable"],"summary":"Get Unpaid List","description":"Get an unpaid payable list\n\n*No parameters*\n\n    Returns a list of every unpaid service (payment_status = pending)","operationId":"get_unpaid_list_unpaid_list__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/unpaid-list-by-service-type/{service_type}":{"get":{"tags":["Payable"],"summary":"Get Unpaid List By Service Type","description":"Get an unpaid payable list by service_type\n\n*Parameter in path*\n\n    service_type: luz, agua, gas, etc","operationId":"get_unpaid_list_by_service_type_unpaid_list_by_service_type__service_type__get","parameters":[{"required":true,"schema":{"title":"Service Type","type":"string"},"name":"service_type","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/pay-tax/":{"put":{"tags":["Transaction"],"summary":"Make Transaction","description":"Make a pay\n\n*Parameters in body*\n\n    pay_method: (e.g. \"credit_card\", \"debit_card\" or \"cash\")\n\n    card_number: 16 digits-length\n\n    amount: float\n\n    barcode: unique, 12 digits-length\n\n    paid_date: (e.g. 2021-01-15). Format: YYYY-MM-DD","operationId":"make_transaction_pay_tax__put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Transaction"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/get-transaction-list/":{"get":{"tags":["Transaction"],"summary":"Get Transaction List","description":"Get a transaction list between 2 dates given\n\n    Returns a list of a date in particular, how many transactions were done that day and the amount accumulated\n\n*Parameters*\n\n    start_date: (e.g. 2021-01-15). Format: YYYY-MM-DD\n    final_date: (e.g. 2021-01-18). Format: YYYY-MM-DD","operationId":"get_transaction_list_get_transaction_list__get","parameters":[{"required":true,"schema":{"title":"Start Date","type":"string","format":"date"},"name":"start_date","in":"query"},{"required":true,"schema":{"title":"Final Date","type":"string","format":"date"},"name":"final_date","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Payable":{"title":"Payable","required":["service_type","description","due_date","amount","payment_status","barcode"],"type":"object","properties":{"service_type":{"title":"Service Type","type":"string"},"description":{"title":"Description","type":"string"},"due_date":{"title":"Due Date","type":"string","format":"date"},"amount":{"title":"Amount","type":"number"},"payment_status":{"title":"Payment Status","type":"string"},"barcode":{"title":"Barcode","type":"integer"}}},"Transaction":{"title":"Transaction","required":["pay_method","amount","barcode","paid_date"],"type":"object","properties":{"pay_method":{"title":"Pay Method","type":"string"},"card_number":{"title":"Card Number","type":"integer"},"amount":{"title":"Amount","type":"number"},"barcode":{"title":"Barcode","type":"integer"},"paid_date":{"title":"Paid Date","type":"string","format":"date"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}},"tags":[{"name":"Payable","description":"Endpoints related to create payables and get unpaid ones"},{"name":"Transaction","description":"Endpoints related to make a transaction"}]}